staging: android: lowmemorykiller: Make the number of adjustment slots visible.
In Ice Cream Sandwich, the lowmemorykiller options are configured by a
java class that is common throughout every device. To work around this
we make the number of oom_adj slots visible as a module parameter. The
java class can then handle this value if it exists, and modify it's
bahaviour if it doesn't (for legacy reasons).